requirements.txt
redis>=5.0.1,<6.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
typing-extensions>=4.5.0,<5.0.0  # 为Python 3.9提供更好的类型提示支持
python-docx==0.8.11
docxtpl>=0.12.0
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
aiofiles>=23.2.1 